debug message

+++ build-all.bat
@@ -1,10 +1,10 @@
 set CURRENT_DIR=%~dp0
 
-rmdir /s /q opentdf-cpp
+if exist opentdf-cpp rmdir /s /q opentdf-cpp
 mkdir opentdf-cpp\lib
 mkdir opentdf-cpp\include
 
-rmdir /s /q build
+if exist build rmdir /s /q build
 mkdir build
 pushd build
